project/requirements.txt
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
Django==3.0.5
django-cors-headers==3.5.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.12.2
djangotoolbox==1.8.0
djongo==1.3.3
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
mongoengine==0.20.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
pymongo==3.11.0
pytz==2020.4
sqlparse==0.2.4
traitlets==5.0.5
wcwidth==0.2.5
